aaa-packet.c: process binary-base64 convert
aaa_bin2base64() says that "The caller is responsible to free the string.", but I don't know to free which string? So I did not do it.
aaa_packet_serialize(): the malloc size of result should change, but i cannot get the new length of base64 string. So I did not do it.
Please be careful when something goes wrong.